requirements.txt
asgiref==3.2.5
astroid==2.2.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==2.1.14
dnspython==1.16.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.1.0
pylint==2.3.1
pymongo==3.10.1
PyMySQL==0.9.3
PySocks==1.7.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.12.0
soupsieve==2.0
sqlparse==0.3.1
typed-ast==1.4.0
wrapt==1.11.2
djangorestframework
django-snippet
pygments